moved generic components to folder
